requirements.txt
litellm
openai>=1.50.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
json-repair>=0.30
anthropic>=0.40.0
